sweets_classfications/sweet_classfication_project/requirement.txt
absl-py==2.1.0
appdirs==1.4.4
asgiref==3.7.2
aspose-cells==24.3.0
astunparse==1.6.3
attr==0.3.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
bleach==5.0.1
boto==2.49.0
boto3==1.34.61
botocore==1.34.61
boxing==0.1.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
djongo==1.3.6
dm-tree==0.1.8
dnspython==2.6.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
expiringdict==1.2.2
flatbuffers==24.3.7
fonttools==4.49.0
gast==0.5.4
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h5py==3.10.0
htmlmin==0.1.12
humansignal-drf-yasg==1.21.9
idna==3.6
ijson==3.2.3
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jsonschema==3.2.0
kaggle==1.6.6
keras==3.0.5
kiwisolver==1.4.5
label-studio==1.11.0
label-studio-converter==0.0.57
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
libclang==16.0.6
lockfile==0.12.2
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
nltk==3.6.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
progressbar==2.5
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
pygoogle-image==1.0.0
pymongo==3.12.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
python-magic-bin==0.4.14
python-slugify==8.0.4
pytz==2022.7.1
PyYAML==6.0.1
redis==3.5.3
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semver==2.13.0
sentry-sdk==1.42.0
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
xmljson==0.2.0
